exercise 5- RabbitMQ/exercise/requirements.txt
pika==1.1.0
python-dateutil==2.8.1
six==1.15.0